requirements.txt
pandas==2.1.4
psycopg2-binary==2.9.9
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
